requirements.txt
praw>=7.7
openai>=1.0
pyyaml>=6.0
rich>=13.0
